Split tests into more classes
